# swagger-new-endpoints.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/settings": {
      "get": {
        "operationId": "SettingsController_get",
        "parameters": [
          {
            "required": false,
            "name": "key",
            "in": "query",
            "schema": {
              "enum": [
                "generalCopyRightTextAr",
                "logo",
                "generalCopyRightTextEn",
                "conditionsAndProvisionsEn",
                "conditionsAndProvisionsAr",
                "privatePolicyAr",
                "privatePolicyEn",
                "shippingKMCharge",
                "deliveryCommission",
                "businessOrderCommissionRateForAll",
                "filterByZone",
                "businessOrderCommissionRate",
                "businessOrderCommissionType",
                "StoreTaxForAll",
                "StoreTaxRate",
                "storeNearestByKM",
                "deliveryAcceptanceTimer",
                "deliveryAssignmentMode",
                "deliveryAfkBreakMinutes",
                "deliveryAfkBreakEnabled",
                "pickupGeofenceRadiusMeters",
                "deliveryGeofenceRadiusMeters",
                "customDeliveryExtraStopPrice",
                "customDeliveryKMCharge",
                "customDeliveryBaseFee",
                "customDeliveryCommissionForAll",
                "customDeliveryCommissionRate",
                "customDeliveryCommissionType",
                "adminDashboardPeriodStartAt"
              ],
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "domain",
            "in": "query",
            "schema": {
              "enum": [
                "BUSINESS",
                "DELIVERY",
                "ORDER",
                "REFUND",
                "STORE",
                "CUSTOMER",
                "PRIORITY",
                "DISBURSEMENT"
              ],
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": { "example": "en" }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [true, false]
            }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Settings"],
        "summary": "Fetch app-wide configuration, filterable by domain/key. Public — no auth required.",
        "description": "Fetch app-wide configuration, filterable by domain/key. Public — no auth required."
      },
      "patch": {
        "operationId": "SettingsController_update",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": { "example": "en" }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [true, false]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "settings": {
                    "type": "string",
                    "example": "[{\"setting\":\"test answer\",\"value\":\"test answer\",\"name\":{}}]"
                  },
                  "settings[0].file": { "type": "string", "format": "binary" },
                  "settings[1].file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Settings"],
        "security": [{ "ACCESS Token": [] }],
        "summary": "Admin updates one or more configuration values, including the delivery geofence radii and custom-delivery pricing settings.",
        "description": "Admin updates one or more configuration values, including the delivery geofence radii and custom-delivery pricing settings."
      }
    },
    "/api/authentication/refresh-token": {
      "post": {
        "operationId": "BaseAuthenticationController_refreshToken",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": { "example": "en" }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [true, false]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": { "default": { "value": { "data": null } } }
              }
            },
            "description": ""
          }
        },
        "tags": ["Authentication"],
        "security": [{ "REFRESH Token": [] }],
        "summary": "Exchanges a refresh token for a new access token. Refresh tokens now actually live for their full configured duration (previously capped at 1 day regardless of setting).",
        "description": "Exchanges a refresh token for a new access token. Refresh tokens now actually live for their full configured duration (previously capped at 1 day regardless of setting)."
      }
    }
  },
  "info": {
    "title": "Makanak — New & Fixed Endpoints",
    "description": "Standalone subset of the full API, containing only the endpoints added or behaviorally changed during this engagement — for isolated testing.",
    "version": "1.0",
    "contact": {
      "name": "Mahmoud Elamrosy",
      "url": "www.linkedin.com/in/",
      "email": "melamrosy72@gmail.com"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "ACCESS Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "REFRESH Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "PASSWORD_RESET Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "VERIFY Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "VISITOR Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}